Separation of the map
